Fix handling of command and args parameter
An empty list for `command` or `args` in the podman RuntimeConfig has
a different meaning than omitting one of the parameters. An empty list
means to overwrite the corresponding setting from the container
image. Omitting one of the parameters means to use the corresponding
default setting of the container image.

Issue-Id: #10