Store stdout and stderr of stests to file
The stdout and stderr output of application used in stests is stored
to temporary files. This is necessary, as storing this output to
memory can lead to test cases fail, if the application hangs as stdout
or stderr are full and the application is not able to output to stdout
or stderr.

Issue-Id: #13