Simplify ank cli set state command (#383)
* Remove unnecessary complexity from ank cli

Before the ank cli, when executing a "set state" command, only send
the parts of the state which have be changed by the update mask. This
was probably done for previous problems with optional fields. This is
now been removed and ank cli sends the whole state provided by the
users, as:

* the implementation did not allow to delete objects from the desired
state
* the ank cli should not have to much knowledge about the
implementation. E.g. if we would introduce wildcards in the mask, this
would needed to be implemented in the server and the ank cli.

* Fix incorrect stest resources

* Add missing fields to stest config