Switch the Control Interface to tokio net pipes (#487)
* Switch the Control Interface to tokio pipes

* Remove unneeded r/w, cleanup and fix  tests

* Avoid busy waiting for the workload to reconnect.
The problem can only occur if unsolicited messages (logs, events) are written, or the workload disconnects exactly at the time of the writing.
The previous implementation using files would have blocked at the opening of the file on the other side, but uses threads for this.

* Refactor old utests to account for new coding styles

* Correct requirement Id and text
We are reading the input and writing to the output.

* Add diagrams for async and sync workload communication